// lib/api/model/modelSalary.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Jeffries Homeapp API
 * Backend REST API for Jeffries Homeapp.
 * OpenAPI spec version: 1.0
 */

export interface ModelSalary {
  aantal_diensten?: number;
  amt_zeerintensief?: number;
  basis_loon?: number;
  berekend_op?: string;
  bruto_betaling?: number;
  eenmalig_totaal?: number;
  extra_uren_bedrag?: number;
  id?: string;
  jaar?: number;
  loonheffing_schat?: number;
  maand?: number;
  netto_prognose?: number;
  ort_totaal?: number;
  pensioenpremie?: number;
  periode?: string;
  reiskosten?: number;
  toeslag_balansvlf?: number;
  toeslag_vakatie_uren?: number;
  user_id?: string;
  uurloon_ort?: number;
}